requirements.txt
transformers
pillow
tensor
torch
torchaudio
streamlit
speechrecognition
openai
sounddevice
numpy
easyocr
pydub